updated according to comments
