+++ ddtrace/tracer/rules_sampler.go
@@ -6,9 +6,14 @@
 package tracer
 
 import (
+	"encoding/json"
+	"fmt"
 	"math"
 	"os"
+	"regexp"
 	"strconv"
+	"strings"
+	"sync"
 	"time"
 
 	"golang.org/x/time/rate"
@@ -17,6 +22,110 @@ import (
 	"gopkg.in/DataDog/dd-trace-go.v1/internal/log"
 	"gopkg.in/DataDog/dd-trace-go.v1/internal/samplernames"
 )
+
+// rulesSampler holds instances of trace sampler and single span sampler, that are configured with the given set of rules.
+// traceRulesSampler samples trace spans based on a user-defined set of rules and might impact sampling decision of the trace.
+// singleSpanRulesSampler samples individual spans based on a separate user-defined set of rules and
+// cannot impact the trace sampling decision.
+type rulesSampler struct {
+	traces *traceRulesSampler
+	spans  *singleSpanRulesSampler
+}
+
+// newRulesSampler configures a *rulesSampler instance using the given set of rules.
+// Rules are split between trace and single span sampling rules according to their type.
+// Such rules are user-defined through environment variable or WithSamplingRules option.
+// Invalid rules or environment variable values are tolerated, by logging warnings and then ignoring them.
+func newRulesSampler(rules []SamplingRule) *rulesSampler {
+	var spanRules, traceRules []SamplingRule
+	for _, rule := range rules {
+		if rule.Type == SamplingRuleSpan {
+			rule.limiter = newSingleSpanRateLimiter(rule.MaxPerSecond)
+			spanRules = append(spanRules, rule)
+		} else {
+			traceRules = append(traceRules, rule)
+		}
+	}
+	return &rulesSampler{
+		traces: newTraceRulesSampler(traceRules),
+		spans:  newSingleSpanRulesSampler(spanRules),
+	}
+}
+
+func (r *rulesSampler) sampleTrace(s *span) bool {
+	return r.traces.apply(s)
+}
+
+func (r *rulesSampler) sampleSpan(s *span) bool {
+	return r.spans.apply(s)
+}
+
+// SamplingRule is used for applying sampling rates to spans that match
+// the service name, operation name or both.
+// For basic usage, consider using the helper functions ServiceRule, NameRule, etc.
+// See `SamplingRuleType` for more details.
+type SamplingRule struct {
+	Service      *regexp.Regexp
+	Name         *regexp.Regexp
+	Rate         float64
+	MaxPerSecond float64
+	Type         SamplingRuleType
+
+	exactService string
+	exactName    string
+	limiter      *rateLimiter
+}
+
+// SamplingRuleType represents a type of sampling rule spans are matched against.
+type SamplingRuleType int
+
+const (
+	// SamplingRuleTrace specifies a sampling rule that applies to the entire trace if any spans satisfy the criteria.
+	// If a sampling rule is of type SamplingRuleTrace, such rule determines the sampling rate to apply
+	// to trace spans. If a span matches that rule, it will impact the trace sampling decision.
+	SamplingRuleTrace = iota
+	// SamplingRuleSpan specifies a sampling rule that applies to a single span without affecting the entire trace.
+	// If a sampling rule is of type SamplingRuleSingleSpan, such rule determines the sampling rate to apply
+	// to individual spans. If a span matches a rule, it will NOT impact the trace sampling decision.
+	// In the case that a trace is dropped and thus not sent to the Agent, spans kept on account
+	// of matching SamplingRuleSingleSpan rules must be conveyed separately.
+	SamplingRuleSpan
+)
+
+// ServiceRule returns a SamplingRule that applies the provided sampling rate
+// to spans that match the service name provided.
+func ServiceRule(service string, rate float64) SamplingRule {
+	return SamplingRule{
+		exactService: service,
+		Rate:         rate,
+	}
+}
+
+// NameRule returns a SamplingRule that applies the provided sampling rate
+// to spans that match the operation name provided.
+func NameRule(name string, rate float64) SamplingRule {
+	return SamplingRule{
+		exactName: name,
+		Rate:      rate,
+	}
+}
+
+// NameServiceRule returns a SamplingRule that applies the provided sampling rate
+// to spans matching both the operation and service names provided.
+func NameServiceRule(name string, service string, rate float64) SamplingRule {
+	return SamplingRule{
+		exactService: service,
+		exactName:    name,
+		Rate:         rate,
+	}
+}
+
+// RateRule returns a SamplingRule that applies the provided sampling rate to all spans.
+func RateRule(rate float64) SamplingRule {
+	return SamplingRule{
+		Rate: rate,
+	}
+}
 
 // traceRulesSampler allows a user-defined list of rules to apply to spans.
 // These rules can match based on the span's Service, Name or both.
@@ -46,6 +155,26 @@ func newTraceRulesSampler(rules []SamplingRule) *traceRulesSampler {
 		globalRate: globalSampleRate(),
 		limiter:    newRateLimiter(),
 	}
+}
+
+// globalSampleRate returns the sampling rate found in the DD_TRACE_SAMPLE_RATE environment variable.
+// If it is invalid or not within the 0-1 range, NaN is returned.
+func globalSampleRate() float64 {
+	defaultRate := math.NaN()
+	v := os.Getenv("DD_TRACE_SAMPLE_RATE")
+	if v == "" {
+		return defaultRate
+	}
+	r, err := strconv.ParseFloat(v, 64)
+	if err != nil {
+		log.Warn("ignoring DD_TRACE_SAMPLE_RATE: error: %v", err)
+		return defaultRate
+	}
+	if r >= 0.0 && r <= 1.0 {
+		return r
+	}
+	log.Warn("ignoring DD_TRACE_SAMPLE_RATE: out of range %f", r)
+	return defaultRate
 }
 
 func (rs *traceRulesSampler) enabled() bool {
@@ -81,6 +210,21 @@ func (rs *traceRulesSampler) apply(span *span) bool {
 	return true
 }
 
+// match returns true when the span's details match all the expected values in the rule.
+func (sr *SamplingRule) match(s *span) bool {
+	if sr.Service != nil && !sr.Service.MatchString(s.Service) {
+		return false
+	} else if sr.exactService != "" && sr.exactService != s.Service {
+		return false
+	}
+	if sr.Name != nil && !sr.Name.MatchString(s.Name) {
+		return false
+	} else if sr.exactName != "" && sr.exactName != s.Name {
+		return false
+	}
+	return true
+}
+
 func (rs *traceRulesSampler) applyRate(span *span, rate float64, now time.Time) {
 	span.SetTag(keyRulesSamplerAppliedRate, rate)
 	if !sampledByRate(span.TraceID, rate) {
@@ -110,16 +254,16 @@ func (rs *traceRulesSampler) limit() (float64, bool) {
 const defaultRateLimit = 100.0
 
 // newRateLimiter returns a rate limiter which restricts the number of traces sampled per second.
-// This defaults to 100.0. The DD_TRACE_RATE_LIMIT environment variable may override the default.
+// The limit is DD_TRACE_RATE_LIMIT if set, `defaultRateLimit` otherwise.
 func newRateLimiter() *rateLimiter {
 	limit := defaultRateLimit
 	v := os.Getenv("DD_TRACE_RATE_LIMIT")
 	if v != "" {
 		l, err := strconv.ParseFloat(v, 64)
 		if err != nil {
-			log.Warn("using default rate limit because DD_TRACE_RATE_LIMIT is invalid: %v", err)
+			log.Warn("DD_TRACE_RATE_LIMIT invalid, using default value %f: %v", limit, err)
 		} else if l < 0.0 {
-			log.Warn("using default rate limit because DD_TRACE_RATE_LIMIT is negative: %f", l)
+			log.Warn("DD_TRACE_RATE_LIMIT negative, using default value %f", limit)
 		} else {
 			// override the default limit
 			limit = l
@@ -129,4 +273,298 @@ func newRateLimiter() *rateLimiter {
 		limiter:  rate.NewLimiter(rate.Limit(limit), int(math.Ceil(limit))),
 		prevTime: time.Now(),
 	}
+}
+
+// singleSpanRulesSampler allows a user-defined list of rules to apply to spans
+// to sample single spans.
+// These rules match based on the span's Service and Name. If empty value is supplied
+// to either Service or Name field, it will default to "*", allow all.
+// When making a sampling decision, the rules are checked in order until
+// a match is found.
+// If a match is found, the rate from that rule is used.
+// If no match is found, no changes or further sampling is applied to the spans.
+// The rate is used to determine if the span should be sampled, but an upper
+// limit can be defined using the max_per_second field when supplying the rule.
+// If max_per_second is absent in the rule, the default is allow all.
+// Its value is the max number of spans to sample per second.
+// Spans that matched the rules but exceeded the rate limit are not sampled.
+type singleSpanRulesSampler struct {
+	rules []SamplingRule // the rules to match spans with
+}
+
+// newSingleSpanRulesSampler configures a *singleSpanRulesSampler instance using the given set of rules.
+// Invalid rules or environment variable values are tolerated, by logging warnings and then ignoring them.
+func newSingleSpanRulesSampler(rules []SamplingRule) *singleSpanRulesSampler {
+	return &singleSpanRulesSampler{
+		rules: rules,
+	}
+}
+
+func (rs *singleSpanRulesSampler) enabled() bool {
+	return len(rs.rules) > 0
+}
+
+const (
+	// spanSamplingMechanism specifies the sampling mechanism by which an individual span was sampled
+	spanSamplingMechanism = "_dd.span_sampling.mechanism"
+
+	// singleSpanSamplingMechanism specifies value reserved to indicate that a span was kept
+	// on account of a single span sampling rule.
+	singleSpanSamplingMechanism = 8
+
+	// singleSpanSamplingRuleRate specifies the configured sampling probability for the single span sampling rule.
+	singleSpanSamplingRuleRate = "_dd.span_sampling.rule_rate"
+
+	// singleSpanSamplingMPS specifies the configured limit for the single span sampling rule
+	// that the span matched. If there is no configured limit, then this tag is omitted.
+	singleSpanSamplingMPS = "_dd.span_sampling.max_per_second"
+)
+
+// apply uses the sampling rules to determine the sampling rate for the
+// provided span. If the rules don't match, then it returns false and the span is not
+// modified.
+func (rs *singleSpanRulesSampler) apply(span *span) bool {
+	for _, rule := range rs.rules {
+		if rule.match(span) {
+			rs.applyRate(span, rule, rule.Rate, time.Now())
+			return true
+		}
+	}
+	return false
+}
+
+func (rs *singleSpanRulesSampler) applyRate(span *span, rule SamplingRule, rate float64, now time.Time) {
+	span.setMetric(keyRulesSamplerAppliedRate, rate)
+	if !sampledByRate(span.SpanID, rate) {
+		span.setSamplingPriority(ext.PriorityUserReject, samplernames.RuleRate, rate)
+		return
+	}
+
+	var sampled bool
+	if rule.limiter != nil {
+		sampled, rate = rule.limiter.allowOne(now)
+		if !sampled {
+			return
+		}
+	}
+	span.setMetric(spanSamplingMechanism, singleSpanSamplingMechanism)
+	span.setMetric(singleSpanSamplingRuleRate, rate)
+	if rule.MaxPerSecond != 0 {
+		span.setMetric(singleSpanSamplingMPS, rule.MaxPerSecond)
+	}
+}
+
+// rateLimiter is a wrapper on top of golang.org/x/time/rate which implements a rate limiter but also
+// returns the effective rate of allowance.
+type rateLimiter struct {
+	limiter *rate.Limiter
+
+	mu          sync.Mutex // guards below fields
+	prevTime    time.Time  // time at which prevAllowed and prevSeen were set
+	allowed     float64    // number of spans allowed in the current period
+	seen        float64    // number of spans seen in the current period
+	prevAllowed float64    // number of spans allowed in the previous period
+	prevSeen    float64    // number of spans seen in the previous period
+}
+
+// allowOne returns the rate limiter's decision to allow the span to be sampled, and the
+// effective rate at the time it is called. The effective rate is computed by averaging the rate
+// for the previous second with the current rate
+func (r *rateLimiter) allowOne(now time.Time) (bool, float64) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
+	if d := now.Sub(r.prevTime); d >= time.Second {
+		// enough time has passed to reset the counters
+		if d.Truncate(time.Second) == time.Second && r.seen > 0 {
+			// exactly one second, so update prev
+			r.prevAllowed = r.allowed
+			r.prevSeen = r.seen
+		} else {
+			// more than one second, so reset previous rate
+			r.prevAllowed = 0
+			r.prevSeen = 0
+		}
+		r.prevTime = now
+		r.allowed = 0
+		r.seen = 0
+	}
+
+	r.seen++
+	var sampled bool
+	if r.limiter.AllowN(now, 1) {
+		r.allowed++
+		sampled = true
+	}
+	er := (r.prevAllowed + r.allowed) / (r.prevSeen + r.seen)
+	return sampled, er
+}
+
+// newSingleSpanRateLimiter returns a rate limiter which restricts the number of single spans sampled per second.
+// This defaults to infinite, allow all behaviour. The MaxPerSecond value of the rule may override the default.
+func newSingleSpanRateLimiter(mps float64) *rateLimiter {
+	limit := math.MaxFloat64
+	if mps > 0 {
+		limit = mps
+	}
+	return &rateLimiter{
+		limiter:  rate.NewLimiter(rate.Limit(limit), int(math.Ceil(limit))),
+		prevTime: time.Now(),
+	}
+}
+
+// globMatch compiles pattern string into glob format, i.e. regular expressions with only '?'
+// and '*' treated as regex metacharacters.
+func globMatch(pattern string) (*regexp.Regexp, error) {
+	// escaping regex characters
+	pattern = regexp.QuoteMeta(pattern)
+	// replacing '?' and '*' with regex characters
+	pattern = strings.Replace(pattern, "\\?", ".", -1)
+	pattern = strings.Replace(pattern, "\\*", ".*", -1)
+	//pattern must match an entire string
+	return regexp.Compile(fmt.Sprintf("^%s$", pattern))
+}
+
+// samplingRulesFromEnv parses sampling rules from the DD_TRACE_SAMPLING_RULES,
+// DD_SPAN_SAMPLING_RULES and DD_SPAN_SAMPLING_RULES_FILE environment variables.
+func samplingRulesFromEnv() ([]SamplingRule, error) {
+	var errs []string
+	var rules []SamplingRule
+	rulesFromEnv := os.Getenv("DD_TRACE_SAMPLING_RULES")
+	if rulesFromEnv != "" {
+		traceRules, err := unmarshalSamplingRules([]byte(rulesFromEnv), SamplingRuleTrace)
+		if err != nil {
+			errs = append(errs, err.Error())
+		}
+		rules = append(rules, traceRules...)
+	}
+	spanRules, err := unmarshalSamplingRules([]byte(os.Getenv("DD_SPAN_SAMPLING_RULES")), SamplingRuleSpan)
+	if err != nil {
+		errs = append(errs, err.Error())
+	}
+	rules = append(rules, spanRules...)
+	rulesFile := os.Getenv("DD_SPAN_SAMPLING_RULES_FILE")
+	if len(rules) != 0 {
+		if rulesFile != "" {
+			log.Warn("DIAGNOSTICS Error(s): DD_SPAN_SAMPLING_RULES is available and will take precedence over DD_SPAN_SAMPLING_RULES_FILE")
+		}
+		if len(errs) != 0 {
+			return rules, fmt.Errorf("%s", strings.Join(errs, "\n\t"))
+		}
+		return rules, nil
+	}
+	if rulesFile != "" {
+		rulesFromEnvFile, err := os.ReadFile(rulesFile)
+		if err != nil {
+			errs = append(errs, fmt.Sprintf("Couldn't read file from DD_SPAN_SAMPLING_RULES_FILE: %v", err))
+		}
+		spanRules, err = unmarshalSamplingRules(rulesFromEnvFile, SamplingRuleSpan)
+		if err != nil {
+			errs = append(errs, err.Error())
+		}
+		rules = append(rules, spanRules...)
+	}
+	if len(errs) != 0 {
+		return rules, fmt.Errorf("%s", strings.Join(errs, "\n\t"))
+	}
+	return rules, nil
+}
+
+// unmarshalSamplingRules unmarshals JSON from b and returns the sampling rules found, attributing
+// the type t to them. If any errors are occurred, they are returned.
+func unmarshalSamplingRules(b []byte, spanType SamplingRuleType) ([]SamplingRule, error) {
+	if len(b) == 0 {
+		return nil, nil
+	}
+	var jsonRules []struct {
+		Service      string      `json:"service"`
+		Name         string      `json:"name"`
+		Rate         json.Number `json:"sample_rate"`
+		MaxPerSecond float64     `json:"max_per_second"`
+	}
+	err := json.Unmarshal(b, &jsonRules)
+	if err != nil {
+		return nil, fmt.Errorf("error unmarshalling JSON: %v", err)
+	}
+	rules := make([]SamplingRule, 0, len(jsonRules))
+	var errs []string
+	for i, v := range jsonRules {
+		if v.Rate == "" {
+			errs = append(errs, fmt.Sprintf("at index %d: rate not provided", i))
+			continue
+		}
+		rate, err := v.Rate.Float64()
+		if err != nil {
+			errs = append(errs, fmt.Sprintf("at index %d: %v", i, err))
+			continue
+		}
+		if !(rate >= 0.0 && rate <= 1.0) {
+			log.Warn("at index %d: ignoring rule %+v: rate is out of [0.0, 1.0] range", i, v)
+			continue
+		}
+		switch spanType {
+		case SamplingRuleSpan:
+			if v.Service == "" {
+				v.Service = "*"
+			}
+			srvGlob, err := globMatch(v.Service)
+			if err != nil {
+				log.Warn("at index %d: ignoring rule %+v: service name regex pattern can't be compiled", i, v)
+				continue
+			}
+			if v.Name == "" {
+				v.Name = "*"
+			}
+			opGlob, err := globMatch(v.Name)
+			if err != nil {
+				log.Warn("at index %d: ignoring rule %+v: operation name regex pattern can't be compiled", i, v)
+				continue
+			}
+			rules = append(rules, SamplingRule{
+				Service:      srvGlob,
+				Name:         opGlob,
+				Rate:         rate,
+				MaxPerSecond: v.MaxPerSecond,
+				limiter:      newSingleSpanRateLimiter(v.MaxPerSecond),
+				Type:         SamplingRuleSpan,
+			})
+		case SamplingRuleTrace:
+			switch {
+			case v.Service != "" && v.Name != "":
+				rules = append(rules, NameServiceRule(v.Name, v.Service, rate))
+			case v.Service != "":
+				rules = append(rules, ServiceRule(v.Service, rate))
+			case v.Name != "":
+				rules = append(rules, NameRule(v.Name, rate))
+			}
+		}
+	}
+	if len(errs) != 0 {
+		return rules, fmt.Errorf("\n\t%s", strings.Join(errs, "\n\t"))
+	}
+	return rules, nil
+}
+
+// MarshalJSON implements the json.Marshaler interface.
+func (sr *SamplingRule) MarshalJSON() ([]byte, error) {
+	s := struct {
+		Service      string   `json:"service"`
+		Name         string   `json:"name"`
+		Rate         float64  `json:"sample_rate"`
+		MaxPerSecond *float64 `json:"max_per_second,omitempty"`
+	}{}
+	if sr.exactService != "" {
+		s.Service = sr.exactService
+	} else if sr.Service != nil {
+		s.Service = fmt.Sprintf("%s", sr.Service)
+	}
+	if sr.exactName != "" {
+		s.Name = sr.exactName
+	} else if sr.Name != nil {
+		s.Name = fmt.Sprintf("%s", sr.Name)
+	}
+	s.Rate = sr.Rate
+	if sr.MaxPerSecond != 0 {
+		s.MaxPerSecond = &sr.MaxPerSecond
+	}
+	return json.Marshal(&s)
 }
