+++ ddtrace/tracer/tracer_test.go
@@ -367,6 +367,7 @@ func TestSamplingDecision(t *testing.T) {
 		assert.Equal(t, "", span.context.trace.propagatingTags[keyDecisionMaker])
 		assert.Equal(t, decisionDrop, span.context.trace.samplingDecision)
 	})
+
 	t.Run("client_dropped_with_single_spans", func(t *testing.T) {
 		os.Setenv("DD_SPAN_SAMPLING_RULES", `[{"service": "test_*","name":"*_1", "sample_rate": 1.0, "max_per_second": 15.0}]`)
 		defer os.Unsetenv("DD_SPAN_SAMPLING_RULES")
